Move declare simd related stuff to its own file
