+++ llvm/lib/Frontend/OpenMP/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_llvm_component_library(LLVMFrontendOpenMP
   OMP.cpp
   OMPContext.cpp
+  OMPDeclareSimd.cpp
   OMPIRBuilder.cpp
   DirectiveNameParser.cpp
